services/tts-piper/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
piper-tts>=1.2.0
pydantic-settings>=2.1.0
